Refactor document view into its own functional component
