+++ src/components/navigation/section-document-or-browser.tsx
@@ -1,24 +1,18 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect } from "react";
 import { observer } from "mobx-react";
 import { useQueryClient } from 'react-query';
-import classNames from "classnames";
 import { DocumentModelType } from "../../models/document/document";
-import { getDocumentDisplayTitle } from "../../models/document/document-utils";
 import { logDocumentEvent } from "../../models/document/log-document-event";
 import { ISubTabSpec, NavTabModelType } from "../../models/view/nav-tabs";
-import { DocumentType } from "src/models/document/document-types";
-import { EditableDocumentContent } from "../document/editable-document-content";
-import { useAppConfig, useClassStore, useLocalDocuments, useProblemStore, useStores,
-  useUIStore, useUserStore } from "../../hooks/use-stores";
+import { useAppConfig, useClassStore, useProblemStore, useStores, useUIStore, useUserStore }
+          from "../../hooks/use-stores";
 import { Logger } from "../../lib/logger";
 import { LogEventName } from "../../lib/logger-types";
 import { useUserContext } from "../../hooks/use-user-context";
 import { NetworkDocumentsSection } from "./network-documents-section";
-import EditIcon from "../../clue/assets/icons/edit-right-icon.svg";
 import { DocumentCollectionList, kNavItemScale } from "../thumbnail/document-collection-list";
 import { SubTabsPanel } from "./sub-tabs-panel";
-import CollapseScrollerIcon from "../../assets/show-hide-document-view-icon.svg";
-import ScrollArrowIcon from "../../assets/scroll-arrow-icon.svg";
+import { DocumentView } from "./document-view";
 
 import "./section-document-or-browser.scss";
 
@@ -37,12 +31,12 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
   const queryClient = useQueryClient();
   const user = useUserStore();
   const classStore = useClassStore();
-  const documents = useLocalDocuments();
   const navTabSpec = appConfigStore.navTabs.getNavTabSpec(tabSpec.tab);
   const subTabs = tabSpec.subTabs;
   const tabState = navTabSpec && ui.tabs.get(navTabSpec?.tab);
   const subTabIndex = Math.max(subTabs.findIndex((subTab) => tabState?.openSubTab === subTab.label), 0);
   const selectedSubTab = subTabs[subTabIndex];
+  // const isStarredTab = selectedSubTab.label === "Starred";
 
   useEffect(() => {
     // Set the initial open tab. If the tabSpec changes somehow then the open
@@ -72,30 +66,9 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
 
   const handleSelectDocument = (document: DocumentModelType) => {
     if (ui.focusDocument === document.key) {
-      if (ui.focusSecondaryDocument) {
-        ui.openSubTabDocument(tabSpec.tab, selectedSubTab.label, ui.focusSecondaryDocument);
-        ui.closeSubTabSecondaryDocument(tabSpec.tab, selectedSubTab.label);
-      } else {
         ui.closeSubTabDocument(tabSpec.tab, selectedSubTab.label);
-      }
-    } else if (selectedSubTab.label === "Starred"){
-      if (tabState?.openDocuments.get("Starred")) {
-        if (ui.focusSecondaryDocument === document.key) {
-          ui.closeSubTabSecondaryDocument(tabSpec.tab, "Starred");
-        } else {
-          ui.openSubTabSecondaryDocument(tabSpec.tab, "Starred", document.key);
-        }
-      } else {
-        if (!document.hasContent && document.isRemote) {
-          loadDocumentContent(document);
-        }
-        ui.openSubTabDocument(tabSpec.tab, selectedSubTab.label, document.key);
-        const logEvent = document.isRemote
-          ? LogEventName.VIEW_SHOW_TEACHER_NETWORK_COMPARISON_DOCUMENT
-          : LogEventName.VIEW_SHOW_COMPARISON_DOCUMENT;
-        logDocumentEvent(logEvent, { document });
-      }
-    } else {
+    }
+    else {
       if (!document.hasContent && document.isRemote) {
         loadDocumentContent(document);
       }
@@ -109,30 +82,6 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
 
   const loadDocumentContent = async (document: DocumentModelType) => {
     await document.fetchRemoteContent(queryClient, context);
-  };
-
-  function handleEditClick(document: DocumentModelType) {
-    ui.problemWorkspace.setPrimaryDocument(document);
-  }
-
-  // TODO: this edit button is confusing when the history is being viewed. It
-  // opens the original document for editing, not some old version of the
-  // document they might be looking at. Previously this edit button was disabled
-  // when the history document was being shown because SectionDocumentOrBrowser
-  // knew the state of playback controls. It no longer knows that state, so now
-  // the edit button is shown all of the time.
-  // PT Story: https://www.pivotaltracker.com/story/show/183416176
-  const editButton = (type: string, sClass: string, document: DocumentModelType) => {
-    return (
-      (type === "my-work") || (type === "learningLog")
-        ?
-          <div className={`edit-button ${sClass}`}
-                onClick={() => handleEditClick(document)}>
-            <EditIcon className={`edit-icon ${sClass}`} />
-            <div>Edit</div>
-          </div>
-        : null
-    );
   };
 
   const renderDocumentBrowserView = (subTab: ISubTabSpec) => {
@@ -162,155 +111,14 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
     );
   };
 
-  const getStarredDocuments = (types: DocumentType[]) => {
-    const docs: any = [];
-      types.forEach((type) => {
-        const docsByTypeArr = documents.byType(type);
-        docsByTypeArr.forEach(doc => {
-          docs.push(doc);
-        });
-      });
-    const visibleDocuments = docs.filter ((doc: { isDeleted: any; }) => !doc.isDeleted);
-    const starredDocs = visibleDocuments.filter((doc: { isStarred: any; }) => doc.isStarred);
-    return starredDocs;
-  };
-
-  //TODO: Need to refactor this if we want to deploy to all tabs
   const renderDocumentView = (subTab: ISubTabSpec) => {
     const openDocumentKey = tabState?.openDocuments.get(subTab.label) || "";
     const openDocument = store.documents.getDocument(openDocumentKey) ||
-      store.networkDocuments.getDocument(openDocumentKey);
-    const openSecondaryDocumentKey = tabState?.openSecondaryDocuments.get(subTab.label) || "";
-    const openSecondaryDocument = store.documents.getDocument(openSecondaryDocumentKey) ||
-      store.networkDocuments.getDocument(openSecondaryDocumentKey);
-    const publishedDoc = openDocument?.type === "publication" || openDocument?.type === "personalPublication"
-                          || openDocument?.type === "learningLogPublication";
-    const showPlayback = user.type && !publishedDoc? appConfigStore.enableHistoryRoles.includes(user.type) : false;
-    const isStarredTab = selectedSubTab.label === "Starred";
-    const documentTypes: DocumentType[] = tabSpec.tab === "class-work"
-                                            ? ["publication"]
-                                            : tabSpec.tab === "my-work" && subTab.label === "Starred"
-                                              ? ["problem", "personal"]
-                                              : [];
-    const starredDocuments = getStarredDocuments(documentTypes);
-    const currentOpenDocIndex = openDocument && starredDocuments.indexOf(openDocument);
-    const currentOpenSecondaryDocIndex = openSecondaryDocument && starredDocuments.indexOf(openSecondaryDocument);
+                            store.networkDocuments.getDocument(openDocumentKey);
+    const isStarredTab = subTab.label === "Starred";
     if (!isStarredTab && (!openDocument || openDocument.getProperty("isDeleted"))) return false;
-    const sectionClass = openDocument?.type === "learningLog" ? "learning-log" : "";
-
-    // Published documents are listed in reverse order of index so previous and next toggles are also reversed
-    const handleShowPrevDocument = (secondary?: boolean) => {
-      let prevDocumentKey = "";
-      const prevDocIndex = currentOpenDocIndex + 1 >= starredDocuments.length ? 0 : currentOpenDocIndex + 1;
-      if (prevDocIndex < starredDocuments.length)
-      { prevDocumentKey = starredDocuments[prevDocIndex].key; }
-      if (secondary) {
-        ui.openSubTabSecondaryDocument(tabSpec.tab, subTab.label, prevDocumentKey);
-      } else {
-        ui.openSubTabDocument(tabSpec.tab, subTab.label, prevDocumentKey);
-      }
-    };
-
-    const handleShowNextDocument = (secondary?: boolean | undefined) => {
-      let nextDocumentKey = "";
-      let nextDocIndex: number;
-      // Workspaces Starred tab show the problem document as the first document in the list
-      // The personal documents are shown in reverse order after the problem document
-      if (tabSpec.tab === "my-work") {
-        if (currentOpenDocIndex === 0) {
-          nextDocIndex = starredDocuments.length - 1;
-        } else {
-          nextDocIndex = currentOpenDocIndex - 1;
-        }
-      } else {
-        nextDocIndex = currentOpenDocIndex - 1;
-      }
-
-      if (nextDocIndex < starredDocuments.length) {
-        nextDocumentKey = starredDocuments[nextDocIndex].key;
-      }
-      if (secondary) {
-        ui.openSubTabSecondaryDocument(tabSpec.tab, subTab.label, nextDocumentKey);
-      } else {
-        ui.openSubTabDocument(tabSpec.tab, subTab.label, nextDocumentKey);
-      }
-    };
-
-    const isLeftFlipperVisible = (tabSpec.tab === "class-work" && (currentOpenDocIndex < starredDocuments.length - 1))
-                                    || (tabSpec.tab !== "class-work" && currentOpenDocIndex > 0);
-    const isRightFlipperVisible = (tabSpec.tab === "class-work" && currentOpenDocIndex > 0)
-                                    || (tabSpec.tab !== "class-work" && (currentOpenDocIndex >= 0 &&
-                                          currentOpenDocIndex !== 1));
-
     return (
-      <div className="scroller-and-document">
-        { isStarredTab &&
-          <DocumentBrowserScroller subTab={subTab} tabSpec={tabSpec} openDocumentKey={openDocumentKey}
-              openSecondaryDocumentKey={openSecondaryDocumentKey} onSelectDocument={handleSelectDocument} />
-        }
-        {(!openDocument || openDocument.getProperty("isDeleted"))
-          ? null
-          : <div className="document-area">
-              <div className={`focus-document ${openSecondaryDocument ? "primary" : ""}`}>
-                <div className={`document-header ${tabSpec.tab} ${sectionClass}
-                                ${openSecondaryDocument ? "primary" : ""}`} onClick={() => ui.setSelectedTile()}>
-                  <div className={`document-title`}>
-                    {getDocumentDisplayTitle(openDocument, appConfigStore, problemStore)}
-                  </div>
-                  {(!openDocument.isRemote)
-                      && editButton(tabSpec.tab, sectionClass, openDocument)}
-                </div>
-                {(isStarredTab && isLeftFlipperVisible) &&
-                  <div className="scroll-arrow-button-wrapper left">
-                    <ScrollButton side={"left"} tab={tabSpec.tab} onScroll={()=>handleShowPrevDocument()}/>
-                  </div>
-                }
-                <EditableDocumentContent
-                  mode={"1-up"}
-                  isPrimary={false}
-                  document={openDocument}
-                  readOnly={true}
-                  showPlayback={showPlayback}
-                />
-                {(isStarredTab && isRightFlipperVisible) &&
-                  <div className="scroll-arrow-button-wrapper right">
-                    <ScrollButton side={"right"} tab={tabSpec.tab} onScroll={()=>handleShowNextDocument()}/>
-                  </div>
-                }
-              </div>
-              {(!openSecondaryDocument || openSecondaryDocument.getProperty("isDeleted"))
-                ? null
-                : <div className="focus-document secondary">
-                    <div className={`document-header ${tabSpec.tab} ${sectionClass} secondary`}
-                          onClick={() => ui.setSelectedTile()}>
-                      <div className={`document-title`}>
-                        {getDocumentDisplayTitle(openSecondaryDocument, appConfigStore, problemStore)}
-                      </div>
-                      {(!openSecondaryDocument.isRemote)
-                          && editButton(tabSpec.tab, sectionClass, openSecondaryDocument)}
-                    </div>
-                    {(isStarredTab && isLeftFlipperVisible) &&
-                      <div className="scroll-arrow-button-wrapper left">
-                        <ScrollButton side={"left"} tab={tabSpec.tab} onScroll={()=>handleShowPrevDocument(true)}/>
-                      </div>
-                    }
-                    <EditableDocumentContent
-                      mode={"1-up"}
-                      isPrimary={false}
-                      document={openSecondaryDocument}
-                      readOnly={true}
-                      showPlayback={showPlayback}
-                    />
-                    {(isStarredTab && isRightFlipperVisible) &&
-                      <div className="scroll-arrow-button-wrapper right">
-                        <ScrollButton side={"right"} tab={tabSpec.tab} onScroll={()=>handleShowNextDocument(true)}/>
-                      </div>
-                    }
-                  </div>
-              }
-            </div>
-        }
-      </div>
+      <DocumentView tabSpec={tabSpec} subTab={subTab} />
     );
   };
 
@@ -324,107 +132,3 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
     />
   );
 });
-
-interface DocumentBrowserScrollerProps {
-  subTab: ISubTabSpec;
-  tabSpec: NavTabModelType;
-  openDocumentKey: string;
-  openSecondaryDocumentKey: string;
-  onSelectDocument: (document: DocumentModelType) => void;
-}
-const DocumentBrowserScroller =
-    ({subTab, tabSpec, openDocumentKey, openSecondaryDocumentKey, onSelectDocument}: DocumentBrowserScrollerProps) => {
-  const [scrollerCollapsed, setScrollerCollapsed] = useState(false);
-  const [collectionElement, setCollectionElement] = useState<HTMLDivElement>();
-  const documentScrollerRef = useRef<HTMLDivElement>(null);
-  const [scrollToLocation, setScrollToLocation] = useState(0);
-  const [panelWidth, setPanelWidth] = useState(0);
-
-  const scrollWidth = collectionElement?.scrollWidth ?? 0;
-
-  useEffect(() => {
-    if(scrollToLocation !== undefined) {
-      collectionElement?.scrollTo({left: scrollToLocation, behavior: "smooth"});
-    }
-  },[collectionElement, scrollToLocation]);
-
-  // Keep track of the size of the containing element
-  useEffect(() => {
-    let obs: ResizeObserver;
-    if (documentScrollerRef.current) {
-      obs = new ResizeObserver(() => {
-        setPanelWidth(documentScrollerRef.current?.clientWidth ?? 0);
-      });
-      obs.observe(documentScrollerRef.current);
-    }
-
-    return () => obs?.disconnect();
-  }, []);
-
-  const handleScrollTo = (side: string) => {
-    const direction = side ==="left" ? -1 : 1;
-    const attemptedScrollTo = scrollToLocation + direction * panelWidth;
-    const scrollTo = Math.max(0, Math.min(scrollWidth - panelWidth, attemptedScrollTo));
-    setScrollToLocation(scrollTo);
-  };
-
-  const handleCollapseScroller = () => {
-    setScrollerCollapsed(!scrollerCollapsed);
-  };
-
-  return (
-    <>
-      <div className={classNames("scroller", {"collapsed": scrollerCollapsed})} ref={documentScrollerRef}>
-        {(scrollToLocation > 0) &&
-            <ScrollEndControl side={"left"} collapsed={scrollerCollapsed} tab={tabSpec.tab}
-                onScroll={handleScrollTo} />
-        }
-        <DocumentCollectionList
-            setCollectionElement={setCollectionElement}
-            subTab={subTab}
-            tabSpec={tabSpec}
-            horizontal={true}
-            collapsed={scrollerCollapsed}
-            selectedDocument={openDocumentKey}
-            selectedSecondaryDocument={openSecondaryDocumentKey}
-            scrollToLocation={scrollToLocation}
-            onSelectDocument={onSelectDocument}
-        />
-        {(scrollToLocation < scrollWidth - panelWidth) &&
-            <ScrollEndControl side={"right"} collapsed={scrollerCollapsed} tab={tabSpec.tab}
-                onScroll={handleScrollTo} />
-        }
-      </div>
-      <div className={classNames("collapse-scroller-button", "themed", tabSpec.tab,
-                {"collapsed": scrollerCollapsed})} onClick={handleCollapseScroller}>
-        <CollapseScrollerIcon className={`scroller-icon ${tabSpec.tab}`}/>
-      </div>
-    </>
-  );
-};
-
-interface IScrollEndControlProps {
-  side: string;
-  collapsed?: boolean;
-  tab: string;
-  onScroll: (side: string) => void
-}
-
-const ScrollEndControl = ({side, collapsed, tab, onScroll}: IScrollEndControlProps) => {
-  return (
-    <div className={classNames("scroller-controls", side, {collapsed})}>
-      <div className={`scroller-controls-overlay ${side}`}/>
-      <ScrollButton side={side} collapsed={collapsed} tab={tab}
-                onScroll={onScroll} />
-    </div>
-  );
-};
-
-const ScrollButton = ({side, collapsed, tab, onScroll}: IScrollEndControlProps) => {
-  return (
-    <div className={classNames("scroll-arrow-button", "themed", tab, side, {collapsed})}
-          onClick={()=>onScroll(side)}>
-      <ScrollArrowIcon className={`scroll-arrow-icon ${side} themed ${tab}`} />
-    </div>
-  );
-};
